requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.12
    # via fsspec
aiosignal==1.3.2
    # via aiohttp
albucore==0.0.23
    # via albumentations
albumentations==2.0.3
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
async-timeout==5.0.1
    # via aiohttp
attrs==25.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
certifi==2025.1.31
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.1
    # via requests
comet-ml==3.49.0
    # via -r requirements.in
configobj==5.0.9
    # via everett
dulwich==0.22.7
    # via comet-ml
everett==3.1.0
    # via comet-ml
filelock==3.17.0
    # via torch
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.2.0
    # via
    #   lightning
    #   pytorch-lightning
    #   torch
hydra-core==1.3.2
    # via -r requirements.in
idna==3.10
    # via
    #   requests
    #   yarl
jinja2==3.1.5
    # via torch
jsonschema==4.23.0
    # via comet-ml
jsonschema-specifications==2024.10.1
    # via jsonschema
lightning==2.5.0.post0
    # via -r requirements.in
lightning-utilities==0.12.0
    # via
    #   lightning
    #   pytorch-lightning
    #   torchmetrics
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
networkx==3.4.2
    # via torch
numpy==2.2.2
    # via
    #   albucore
    #   albumentations
    #   opencv-python-headless
    #   scipy
    #   torchmetrics
    #   torchvision
nvidia-cublas-cu12==12.4.5.8
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via torch
nvidia-cuda-runtime-cu12==12.4.127
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.2.1.3
    # via torch
nvidia-curand-cu12==10.3.5.147
    # via torch
nvidia-cusolver-cu12==11.6.1.9
    # via torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.2
    # via torch
nvidia-nccl-cu12==2.21.5
    # via torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via torch
omegaconf==2.3.0
    # via hydra-core
opencv-python-headless==4.11.0.86
    # via
    #   albucore
    #   albumentations
packaging==24.2
    # via
    #   hydra-core
    #   lightning
    #   lightning-utilities
    #   pytorch-lightning
    #   torchmetrics
pillow==11.1.0
    # via torchvision
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
psutil==6.1.1
    # via comet-ml
pydantic==2.10.6
    # via albumentations
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
python-box==6.1.0
    # via comet-ml
pytorch-lightning==2.5.0.post0
    # via lightning
pyyaml==6.0.2
    # via
    #   albumentations
    #   lightning
    #   omegaconf
    #   pytorch-lightning
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   comet-ml
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via comet-ml
rich==13.9.4
    # via comet-ml
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
scipy==1.15.1
    # via albumentations
semantic-version==2.10.0
    # via comet-ml
sentry-sdk==2.20.0
    # via comet-ml
setuptools==75.8.0
    # via lightning-utilities
simplejson==3.19.3
    # via comet-ml
simsimd==6.2.1
    # via albucore
stringzilla==3.11.3
    # via albucore
sympy==1.13.1
    # via torch
torch==2.6.0
    # via
    #   -r requirements.in
    #   lightning
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
torchmetrics==1.6.1
    # via
    #   lightning
    #   pytorch-lightning
torchvision==0.21.0
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   lightning
    #   pytorch-lightning
triton==3.2.0
    # via torch
typing-extensions==4.12.2
    # via
    #   lightning
    #   lightning-utilities
    #   multidict
    #   pydantic
    #   pydantic-core
    #   pytorch-lightning
    #   referencing
    #   rich
    #   torch
urllib3==2.3.0
    # via
    #   comet-ml
    #   dulwich
    #   requests
    #   sentry-sdk
wrapt==1.17.2
    # via comet-ml
wurlitzer==3.1.1
    # via comet-ml
yarl==1.18.3
    # via aiohttp
